QWindow: move context menu synthesis code into private virtual
QQuickWindow needs to be able to call this to ensure that it gets
the context menu event after the mouse event, and not the other way
around, otherwise the menu is immediately closed after opening.

Move the code into QWindowPrivate::maybeSynthesizeContextMenuEvent(),
which is called only for mouse press and release events, and which does
the synthesis only if the event was not already accepted and has no
exclusive grabber.  Use scenePosition() to avoid getting a localized
position left over from delivery to specific widgets or Qt Quick items.
Add explanations to internal docs.

This also opens up the opportunity for QQuickWindowPrivate to do
this in a Qt Quick-specific way.

Task-number: QTBUG-67331
Task-number: QTBUG-93486
Change-Id: I909671d9d62c9007b22646cbea6eede7465ab686
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>