Baseline tests: disable animations in the vista style
The Vista style uses a dynamic property on each objects to control
animations; respect that dynamic property on the style itself as well
to make it possible to turn off animations.

Do so in the baseline test framework to get rid of some flakiness when
using the Windows Vista style.

Change-Id: I2bad329786f38e37d99f947f121e5cd44f5741c1
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Wladimir Leuschner <wladimir.leuschner@qt.io>