tracegen: Remove the use of QUuid and therefore QCryptographicHash
By bypassing the two completely and just using the SHA1 algorithm
directly. This removes the only use of either class in bootstrapped
tools.

The use of the null UUID as a namespace is non-standard, so I've
scheduled it for removal in Qt 7, which will also simplify the code. It
will be closer to Microsoft's suggestion in [1]. Until then, this keeps
the exact same algorithm as Qt 6.8's tracegen so the two versions
generate the same IDs.

[1] https://learn.microsoft.com/en-us/windows/win32/api/traceloggingprovider/nf-traceloggingprovider-tracelogging_define_provider#provider-name-and-id

Change-Id: I5056b93971c9b03d7d82fffd53bf27228074d9f4
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>