Bootstrap: disable QCborValue's use of QCborStreamWriter
We do use QCborValue and QCborStreamWriter in the bootstrap lib, but we
don't need QCborValue's serialization functionality. In fact, we only
need QCborValue because it is used as a backend to QJsonValue.

Drive-by mark the one out-streaming QDataStream method as depending on
QT_CONFIG(cborstreamreader) (which indicates no one is disabling that
feature).

Change-Id: I3e3a3baa0ce28b744adafffdf7dbf0ffe2911208
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>