Android: simplify orientation change handling
Currently we rely on both onConfigurationChanged() and onSizeChanged()
in QtRootLayout to report orientation changes, because the former might
be invoked when the new size is not updated yet, so we're doing many
calculations to ensure we don't end up missing or calling change handler
more than we should.

However, instead of all that, we could simply post
the handler under onConfigurationChanged() which would end up eventually
invoked after onSizeChanged() has been called already.

Task-number: QTBUG-132718
Change-Id: I9fbc826518b81a9e4417eba6cd3cf999637201b9
Reviewed-by: Petri Virkkunen <petri.virkkunen@qt.io>